Добалвяет в компоненты метод для установки обработчиков
