+++ src/components/film-content/film-extra-rated.js
@@ -1,5 +1,5 @@
 import {createElement, createElements, render, RenderPosition} from "../../utils/render";
-import {conditionalTemplate} from "../../util";
+import {conditionalTemplate} from "../../utils/common";
 import {FILM_EXTRA_COUNT} from "../../const";
 import FilmCardComponent from "./film-card";
 import AbstractComponent from "../abstract-component";
