+++ src/components/film-content/show-more-button.js
@@ -7,4 +7,8 @@ export default class ShowMoreButtonComponent extends AbstractComponent {
       `<button class="films-list__show-more">Show more</button>`
     );
   }
+
+  setClickHandler(handler) {
+    this.getElement().addEventListener(`click`, handler);
+  }
 }
